split out ndindex
